consolidate type inference
